cleanup on aisle 7
